requirements.txt
PySide6
pynput
qtawesome
pywin32
opencv-python
numpy
Pillow